+++ packages/common/src/types.ts
@@ -459,6 +459,10 @@ export interface ChainParametersCommon {
     foundationAccount: Base58String;
     /** The chain foundation account index */
     foundationAccountIndex?: bigint;
+    /** Keys allowed to do level1 updates */
+    level1Keys: KeysWithThreshold;
+    /** Keys allowed to do root updates */
+    rootKeys: KeysWithThreshold;
 }
 
 /** Chain parameters used from protocol version 1-3 */
@@ -469,6 +473,8 @@ export type ChainParametersV0 = ChainParametersCommon &
         electionDifficulty: number;
         /** The election difficulty for consensus lottery */
         rewardParameters: RewardParametersV0;
+        /** Keys allowed to do parameter updates */
+        level2Keys: AuthorizationsV0;
     };
 
 /** Chain parameters used in protocol versions 4 and 5 */
@@ -480,6 +486,8 @@ export type ChainParametersV1 = ChainParametersCommon &
         electionDifficulty: number;
         /** The election difficulty for consensus lottery */
         rewardParameters: RewardParametersV1;
+        /** Keys allowed to do parameter updates */
+        level2Keys: AuthorizationsV1;
     };
 
 /** Chain parameters used from protocol version 6 */
@@ -492,6 +500,8 @@ export type ChainParametersV2 = ChainParametersCommon &
     ConsensusParameters & {
         /** The election difficulty for consensus lottery */
         rewardParameters: RewardParametersV2;
+        /** Keys allowed to do parameter updates */
+        level2Keys: AuthorizationsV1;
     };
 
 /** Union of all chain parameters across all protocol versions */
@@ -639,7 +649,10 @@ interface UpdatesCommon {
  * @deprecated This is type describing return types from the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
  */
 export interface UpdatesV0 extends UpdatesCommon {
-    chainParameters: ChainParametersV0;
+    chainParameters: Omit<
+        ChainParametersV0,
+        'level1Keys' | 'level2Keys' | 'rootKeys'
+    >;
     updateQueues: UpdateQueuesV0;
     keys: KeysV0;
 }
@@ -649,7 +662,10 @@ export interface UpdatesV0 extends UpdatesCommon {
  * @deprecated This is type describing return types from the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
  */
 export interface UpdatesV1 extends UpdatesCommon {
-    chainParameters: ChainParametersV1;
+    chainParameters: Omit<
+        ChainParametersV1,
+        'level1Keys' | 'level2Keys' | 'rootKeys'
+    >;
     updateQueues: UpdateQueuesV1;
     keys: KeysV1;
 }
@@ -659,7 +675,10 @@ export interface UpdatesV1 extends UpdatesCommon {
  * @deprecated This is type describing return types from the JSON-RPC client and the V1 gRPC client, both of which have been deprecated
  */
 export interface UpdatesV2 extends UpdatesCommon {
-    chainParameters: ChainParametersV2;
+    chainParameters: Omit<
+        ChainParametersV2,
+        'level1Keys' | 'level2Keys' | 'rootKeys'
+    >;
     updateQueues: UpdateQueuesV2;
     keys: KeysV1;
 }
@@ -2028,3 +2047,12 @@ export type SmartContractTypeValues =
     | number
     | string
     | boolean;
+
+export type HealthCheckResponse =
+    | {
+          isHealthy: true;
+      }
+    | {
+          isHealthy: false;
+          message?: string;
+      };